[BUGFIX] Meaningful message on unreadable file
When the CLI import is unable to open the requested file, there is no
point in just throwing the t3lib_div::getURL() result (which is FALSE)
at the XML parser and then throw its error message

  Badly formatted file (XML2Tree parsing error: Line 1: Invalid document
  end)

at the user. Instead, we want to detect the file read error and provide
the user with a more helpful error message.